manifest: move template to dedicated subpackage
